docs(migration-timeline): refactor to file by feature
