+++ docs_app/src/app/custom-elements/migration-timeline/data-access/migration-timeline-struckture/migration-timeline.data.ts
@@ -1,4 +1,6 @@
-import {MigrationReleaseItem, SubjectSymbols} from './migration-timeline-struckture/interfaces';
+
+import {MigrationReleaseItem} from './migration-item';
+import {SubjectSymbols} from './migration-uid';
 
 export const deprecationAndBreakingChangeTimeline: MigrationReleaseItem[] = [
   {
