--- docs_app/src/app/custom-elements/migration-timeline/data-access/migration-timeline.interface.ts
@@ -1,27 +1,0 @@
-import {semVerString, Deprecation, BreakingChange} from './migration-timeline-struckture/interfaces';
-
-export interface MigrationItemUIDAware {
-  migrationItemUID: string;
-  migrationReleaseUID: string;
-  migrationItemSubjectUID: string;
-  opponentMigrationItemUID: string;
-}
-
-export interface ClientDeprecation extends Deprecation, MigrationItemUIDAware {
-
-}
-
-export interface ClientBreakingChange extends BreakingChange, MigrationItemUIDAware {
-
-}
-
-export interface ClientMigrationTimelineReleaseItem {
-  version: semVerString;
-  // semverNumber 0 0 0 000 0,
-  versionNumber: number;
-  officialRelease: boolean;
-  // JS Date
-  date: Date;
-  deprecations: ClientDeprecation[];
-  breakingChanges: ClientBreakingChange[];
-}
