+++ docs_app/src/app/custom-elements/migration-timeline/data-access/migration-timeline.service.ts
@@ -1,11 +1,10 @@
 import {Injectable} from '@angular/core';
 import {Observable, of, pipe} from 'rxjs';
 import {map, startWith} from 'rxjs/operators';
-import {formatSemVerNumber, parseMigrationItemSubjectUID, parseMigrationItemUID} from '../utils/formatter-parser';
+
 import {LocalState} from '../utils/local-state.service';
-import {MigrationReleaseItem} from './migration-timeline-struckture/interfaces';
-import {deprecationAndBreakingChangeTimeline} from './migration-timeline.data';
-import {ClientMigrationTimelineReleaseItem, MigrationItemUIDAware} from './migration-timeline.interface';
+import {ClientMigrationTimelineReleaseItem, compareByReleaseDateAsc, parseClientMigrationTimelineReleaseItem} from './migration-item';
+import {deprecationAndBreakingChangeTimeline} from './migration-timeline-struckture/migration-timeline.data';
 
 export interface MigrationTimelineState {
   migrations: ClientMigrationTimelineReleaseItem[]
@@ -15,80 +14,23 @@ export interface MigrationTimelineState {
 export class MigrationTimelineService extends LocalState<MigrationTimelineState> {
 
   private initialMigrations: ClientMigrationTimelineReleaseItem[] = [];
-  private staticMigrations: ClientMigrationTimelineReleaseItem[] = this.
-  _parseClientMigrationTimelineReleaseList(deprecationAndBreakingChangeTimeline);
+  private staticMigrations: ClientMigrationTimelineReleaseItem[] =
+    deprecationAndBreakingChangeTimeline.map(release => parseClientMigrationTimelineReleaseItem(release));
   migrations$: Observable<ClientMigrationTimelineReleaseItem[]> = this.select(pipe(
     map(s => s.migrations), startWith(this.initialMigrations))
   );
 
   constructor() {
     super();
+    this.fetchMigrationTimeline();
   }
 
   fetchMigrationTimeline(): void {
     this.connectState(of({
-        migrations: this.staticMigrations.sort((a, b) => {
-          if (a.date < b.date) {
-            return -1;
-          }
-          if (a.date > b.date) {
-            return 1;
-          }
-          return 0;
-        })
+        // ensure base sorting by date before putting it into client state
+        migrations: this.staticMigrations.sort(compareByReleaseDateAsc)
       })
     );
-  }
-
-
-  private _parseClientMigrationTimelineReleaseList(list: MigrationReleaseItem[]): ClientMigrationTimelineReleaseItem[] {
-    return list.map(release => this._parseClientMigrationTimelineReleaseItem(release));
-  }
-
-  private _parseClientMigrationTimelineReleaseItem(r: MigrationReleaseItem): ClientMigrationTimelineReleaseItem {
-    return {
-      ...r,
-      // @TODO create date obj for GMT+0 (or figure out in which timezone RxJS is released...)
-      // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/parse
-      // new Date(var + 'GMT+0')
-      date: new Date(r.date),
-      version: r.version,
-      officialRelease: r.version.split('-').length < 2,
-      versionNumber: formatSemVerNumber(r.version),
-      deprecations: r.deprecations
-        .map(this._parseToMigrationItemUIDAware(r.version)),
-      breakingChanges: r.breakingChanges
-        .map(this._parseToMigrationItemUIDAware(r.version)),
-    } as ClientMigrationTimelineReleaseItem;
-  }
-
-  private _parseToMigrationItemUIDAware<T>(version: string) {
-    return (item: any) => {
-        const awareItem: T & MigrationItemUIDAware = {
-          ...item,
-          migrationReleaseUID: formatSemVerNumber(version) + '',
-          migrationItemSubjectUID: parseMigrationItemSubjectUID(item, {}),
-          migrationItemUID: parseMigrationItemUID(item, {version})
-        };
-
-        let opponentMigrationItemUID;
-        if (item.itemType === 'deprecation') {
-          opponentMigrationItemUID = parseMigrationItemUID(item, {
-            itemType: 'breakingChange',
-            version: item.breakingChangeVersion,
-            subjectAction: item.breakingChangeSubjectAction
-          });
-          // i.itemType === 'breakingChange'
-        } else {
-          opponentMigrationItemUID = parseMigrationItemUID(item, {
-            itemType: 'deprecation',
-            version: item.deprecationVersion,
-            subjectAction: item.deprecationSubjectAction
-          });
-        }
-        awareItem.opponentMigrationItemUID = opponentMigrationItemUID;
-        return awareItem;
-    }
   }
 
 }
